backend/requirements.txt
# Django Core - tối thiểu
Django==5.2.8

# Django REST Framework cho API
djangorestframework==3.15.2

# CORS Headers để frontend React có thể gọi API
django-cors-headers==4.3.1

# Authentication & JWT
djangorestframework-simplejwt==5.5.1

# Environment variables
python-dotenv==1.0.1

# Validation & Serialization
django-filter==24.3

# Password validation
argon2-cffi==23.1.0

# Timezone support
pytz==2024.2

# Pillow cho image upload (nếu cần)
Pillow==11.0.0

# Python dateutil
python-dateutil==2.9.0

# WSGI server for production
gunicorn==21.2.0

# MongoDB Atlas
mongoengine==0.28.2
pymongo==4.6.0
dnspython==2.4.2
djongo

# PyJWT
PyJWT==2.8.0

# Swagger UI
drf-yasg==1.21.7

# Machine Learning
# Lưu ý: Các package này đã được cài sẵn trong Python 3.13
# numpy 2.1.1, scikit-learn 1.7.1, pandas 2.2.3, joblib 1.5.1
# Nếu chưa có, cài riêng: pip install numpy scikit-learn pandas joblib
# Không cần thêm vào requirements.txt vì đã có sẵn